+++ CHANGELOG.md
@@ -8,6 +8,11 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
 
 - Fixed the hook scope in the disguiser causing an error
 - Fixed the classic entity spawn mode breaking on maps without all three spawn types
+- Fixed weapons not using their average firerate with a tickrate dependent fix. Function `SWEP:SetNextPrimaryFire(nextTime)` was overwritten with our fix `SWEP:SetNextPrimaryFire(nextTime, skipTickrateFix)`
+
+### Changed
+
+- Added new param `skipTickrateFix` to `SWEP:SetNextPrimaryFire(nextTime, skipTickrateFix)` to skip our inbuilt tickrate fix
 
 ## [v0.10.2b](https://github.com/TTT-2/TTT2/tree/v0.10.2b) (2021-10-21)