+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
 
 ## Unreleased
 
+## [v0.10.3b](https://github.com/TTT-2/TTT2/tree/v0.10.3b) (2021-10-29)
+
 ### Fixed
 
 - Fixed the hook scope in the disguiser causing an error